+++ Sources/Queuer/Queuer.swift
@@ -31,20 +31,20 @@ public class Queuer {
     /// Shared Queuer.
     public static let shared = Queuer(name: "Queuer")
     
-    /// Queuer OperationQueue.
+    /// Queuer `OperationQueue`.
     public let queue = OperationQueue()
     
-    /// Total Operation count in queue.
+    /// Total `Operation` count in queue.
     public var operationCount: Int {
         return queue.operationCount
     }
     
-    /// Operations currently in queue.
+    /// `Operation`s currently in queue.
     public var operations: [Operation] {
         return queue.operations
     }
     
-    /// The default service level to apply to operations executed using the queue.
+    /// The default service level to apply to `Operation`s executed using the queue.
     public var qualityOfService: QualityOfService {
         get {
             return queue.qualityOfService
@@ -61,7 +61,7 @@ public class Queuer {
         return !queue.isSuspended
     }
     
-    /// Define the max concurrent operation count.
+    /// Define the max concurrent `Operation`s count.
     public var maxConcurrentOperationCount: Int {
         get {
             return queue.maxConcurrentOperationCount
@@ -75,37 +75,75 @@ public class Queuer {
     ///
     /// - Parameters:
     ///   - name: Custom queue name.
-    ///   - maxConcurrentOperationCount: The max concurrent operation count.
-    ///   - qualityOfService: The default service level to apply to operations executed using the queue.
+    ///   - maxConcurrentOperationCount: The max concurrent `Operation`s count.
+    ///   - qualityOfService: The default service level to apply to `Operation`s executed using the queue.
     public init(name: String, maxConcurrentOperationCount: Int = Int.max, qualityOfService: QualityOfService = .default) {
         queue.name = name
         self.maxConcurrentOperationCount = maxConcurrentOperationCount
         self.qualityOfService = qualityOfService
     }
     
-    /// Add an Operation to be executed asynchronously.
+    /// Cancel all `Operation`s in queue.
+    public func cancelAll() {
+        queue.cancelAllOperations()
+    }
+    
+    /// Pause the queue.
+    public func pause() {
+        queue.isSuspended = true
+        
+        for operation in queue.operations {
+            if let concurrentOperation = operation as? ConcurrentOperation {
+                concurrentOperation.pause()
+            }
+        }
+    }
+    
+    /// Resume the queue.
+    public func resume() {
+        queue.isSuspended = false
+        
+        for operation in queue.operations {
+            if let concurrentOperation = operation as? ConcurrentOperation {
+                concurrentOperation.resume()
+            }
+        }
+    }
+    
+    /// Blocks the current thread until all of the receiver’s queued and executing
+    /// `Operation`s finish executing.
+    public func waitUntilAllOperationsAreFinished() {
+        queue.waitUntilAllOperationsAreFinished()
+    }
+}
+
+// MARK: - Queuer Operations and Chaining
+
+/// `Queuer` extension with `Operation`s and chaining handling.
+public extension Queuer {
+    /// Add an `Operation` to be executed asynchronously.
     ///
     /// - Parameter block: Block to be executed.
     public func addOperation(_ operation: @escaping () -> Void) {
         queue.addOperation(operation)
     }
     
-    /// Add an Operation to be executed asynchronously.
+    /// Add an `Operation` to be executed asynchronously.
     ///
-    /// - Parameter operation: Operation to be executed.
+    /// - Parameter operation: `Operation` to be executed.
     public func addOperation(_ operation: Operation) {
         queue.addOperation(operation)
     }
     
-    /// Add an Array of chained Operations.
+    /// Add an Array of chained `Operation`s.
     ///
     /// Example:
     ///
     ///     [A, B, C] = A -> B -> C -> completionHandler.
     ///
     /// - Parameters:
-    ///   - operations: Operations Array.
-    ///   - completionHandler: Completion block to be exectuted when all Operations
+    ///   - operations: `Operation`s Array.
+    ///   - completionHandler: Completion block to be exectuted when all `Operation`s
     ///                        are finished.
     public func addChainedOperations(_ operations: [Operation], completionHandler: (() -> Void)? = nil) {
         for (index, operation) in operations.enumerated() {
@@ -120,53 +158,62 @@ public class Queuer {
             return
         }
         
-        let completionOperation = BlockOperation(block: completionHandler)
-        if !operations.isEmpty {
-            completionOperation.addDependency(operations[operations.count - 1])
-        }
-        addOperation(completionOperation)
+        addCompletionHandler(completionHandler)
     }
     
-    /// Add an Array of chained Operations.
+    /// Add an Array of chained `Operation`s.
     ///
     /// Example:
     ///
     ///     [A, B, C] = A -> B -> C -> completionHandler.
     ///
     /// - Parameters:
-    ///   - operations: Operations list.
-    ///   - completionHandler: Completion block to be exectuted when all Operations
+    ///   - operations: `Operation`s list.
+    ///   - completionHandler: Completion block to be exectuted when all `Operation`s
     ///                        are finished.
     public func addChainedOperations(_ operations: Operation..., completionHandler: (() -> Void)? = nil) {
         addChainedOperations(operations, completionHandler: completionHandler)
     }
     
-    /// Cancel all Operations in queue.
-    public func cancelAll() {
-        queue.cancelAllOperations()
-    }
-    
-    /// Pause the queue.
-    public func pause() {
-        queue.isSuspended = true
-        
-        for operation in queue.operations where operation is ConcurrentOperation {
-            (operation as? ConcurrentOperation)?.pause()
+    /// Add a completion block to the queue.
+    ///
+    /// - Parameter completionHandler: Completion handler to be executed as last `Operation`.
+    public func addCompletionHandler(_ completionHandler: @escaping () -> Void) {
+        let completionOperation = BlockOperation(block: completionHandler)
+        if let lastOperation = operations.last {
+            completionOperation.addDependency(lastOperation)
         }
+        addOperation(completionOperation)
+    }
+}
+
+// MARK: - Queue State Restoration
+
+/// `Queuer` extension with state restoration feature.
+public extension Queuer {
+    /// `OperationState` array typealias.
+    public typealias QueueStateList = [OperationState]
+    
+    /// Creates the queue state.
+    ///
+    /// - Returns: Returns the current queue state.
+    public func state() -> QueueStateList {
+        return Queuer.state(of: queue)
     }
     
-    /// Resume the queue.
-    public func resume() {
-        queue.isSuspended = false
+    /// Creates the state of a given queue.
+    ///
+    /// - Parameter queue: State will be created starting from this `OperationQueue`.
+    /// - Returns: Returns the current queue state.
+    public static func state(of queue: OperationQueue) -> QueueStateList {
+        var operations: QueueStateList = []
         
-        for operation in queue.operations where operation is ConcurrentOperation {
-            (operation as? ConcurrentOperation)?.resume()
+        for operation in queue.operations {
+            if let concurrentOperation = operation as? ConcurrentOperation, let operationName = concurrentOperation.name {
+                operations.append(OperationState(name: operationName, progress: concurrentOperation.progress, dependencies: operation.dependencies.compactMap { $0.name }))
+            }
         }
-    }
-    
-    /// Blocks the current thread until all of the receiver’s queued and executing
-    /// operations finish executing.
-    public func waitUntilAllOperationsAreFinished() {
-        queue.waitUntilAllOperationsAreFinished()
+        
+        return operations
     }
 }