+++ Tests/QueuerTests/SynchronousOperationTests.swift
@@ -29,21 +29,15 @@ import Dispatch
 import XCTest
 
 internal class SynchronousOperationTests: XCTestCase {
-    internal static let allTests = [
-        ("testSynchronousOperation", testSynchronousOperation),
-        ("testSynchronousOperationOnSharedQueuer", testSynchronousOperationOnSharedQueuer),
-        ("testCancel", testCancel)
-    ]
-    
     internal func testSynchronousOperation() {
         let queue = Queuer(name: "SynchronousOperationTestSynchronousOperation")
         let testExpectation = expectation(description: "Synchronous Operation")
         var testString = ""
         
-        let synchronousOperation1 = SynchronousOperation {
+        let synchronousOperation1 = SynchronousOperation { _ in
             testString = "Tested1"
         }
-        let synchronousOperation2 = SynchronousOperation {
+        let synchronousOperation2 = SynchronousOperation { _ in
             Thread.sleep(forTimeInterval: 2)
             testString = "Tested2"
             
@@ -65,10 +59,10 @@ internal class SynchronousOperationTests: XCTestCase {
         let testExpectation = expectation(description: "Synchronous Operation")
         var testString = ""
         
-        let synchronousOperation1 = SynchronousOperation {
+        let synchronousOperation1 = SynchronousOperation { _ in
             testString = "Tested1"
         }
-        let synchronousOperation2 = SynchronousOperation {
+        let synchronousOperation2 = SynchronousOperation { _ in
             Thread.sleep(forTimeInterval: 2)
             testString = "Tested2"
             
@@ -86,6 +80,35 @@ internal class SynchronousOperationTests: XCTestCase {
         }
     }
     
+    internal func testSynchronousOperationRetry() {
+        let queue = Queuer(name: "SynchronousOperationTestRetry")
+        let testExpectation = expectation(description: "Synchronous Operation Retry")
+        var order: [Int] = []
+        
+        let synchronousOperation1 = SynchronousOperation { operation in
+            Thread.sleep(forTimeInterval: 1)
+            order.append(0)
+            operation.hasFailed = true
+            
+            if operation.currentAttempt == 3 {
+                testExpectation.fulfill()
+            }
+        }
+        let synchronousOperation2 = SynchronousOperation { _ in
+            order.append(1)
+        }
+        synchronousOperation1.addToQueue(queue)
+        synchronousOperation2.addToQueue(queue)
+        
+        XCTAssertFalse(synchronousOperation1.isAsynchronous)
+        XCTAssertFalse(synchronousOperation2.isAsynchronous)
+        
+        waitForExpectations(timeout: 5) { error in
+            XCTAssertNil(error)
+            XCTAssertEqual(order, [1, 0, 0, 0])
+        }
+    }
+    
     internal func testCancel() {
         let queue = Queuer(name: "SynchronousOperationTestCancel")
         queue.maxConcurrentOperationCount = 1
@@ -98,11 +121,11 @@ internal class SynchronousOperationTests: XCTestCase {
             testExpectation.fulfill()
         }
         
-        let synchronousOperation1 = SynchronousOperation {
+        let synchronousOperation1 = SynchronousOperation { _ in
             testString = "Tested1"
             Thread.sleep(forTimeInterval: 4)
         }
-        let synchronousOperation2 = SynchronousOperation {
+        let synchronousOperation2 = SynchronousOperation { _ in
             testString = "Tested2"
         }
         synchronousOperation1.addToQueue(queue)