+++ CHANGELOG.md
@@ -8,6 +8,9 @@ All notable changes to this project will be documented in this file.<br>
 
 ---
 
+### 2.x Releases
+- `2.0.x` Releases - [2.0.0](#200---let-me-retry)
+
 ### 1.x Releases
 - `1.3.x` Releases - [1.3.0](#130---open-everything) | [1.3.1](#131---swift-41-support) | [1.3.2](#132---linux-quality)
 - `1.2.x` Releases - [1.2.0](#120---swift-4-support) | [1.2.1](#121---unwanted-alert)
@@ -16,12 +19,34 @@ All notable changes to this project will be documented in this file.<br>
 
 ---
 
-## Develop
+## [2.0.0](https://github.com/FabrizioBrancati/Queuer/releases/tag/2.0.0) - Let Me Retry
+### XX XXX 2018
+### Added
+- Added support to Xcode 10 and Swift 4.2
+- Added retry feature to `ConcurrentOperation` class [#10](https://github.com/FabrizioBrancati/BFKit-Swift/issues/10), more info on how to use it [here](https://github.com/FabrizioBrancati/Queuer#automatically-retry-an-operation) and [here](https://github.com/FabrizioBrancati/Queuer#manually-retry-an-operation)
+- Added `addCompletionHandler(_:)` function to `Queuer` class
+- Added a `Scheduler` class to better schedule your tasks, more info on how to use it [here](https://github.com/FabrizioBrancati/Queuer#scheduler)
+- Added queue state restoration (beta) feature, more info on how to use it [here](https://github.com/FabrizioBrancati/Queuer#queue-state-restoration-beta)
+
+### Changed
+- Changed watchOS target to watchOS 3.0 instead of 2.0, thanks to an App Store submission rule [#11](https://github.com/FabrizioBrancati/BFKit-Swift/issues/11)
+- Changed `executionBlock` of `ConcurrentOperation` to pass the `concurrentOperation` variable inside the closure to be able to use the retry feature. If you don't need it simply put `_ in` after the block creation:
+  ```swift
+  let concurrentOperation = ConcurrentOperation { _ in
+      /// Your task here
+  }
+  ```
+  This also affects `SynchronousOperation`
+- Changed from Codecov to Coveralls service for code coverage
+
 ### Improved
+- Improved `Semaphore` with timeout handling
 - Updated SwiftLint to 0.27.0
 
 ### Removed
 - Removed Hound CI
+
+Thanks to [@zykloman](https://github.com/zykloman) and [@debjitk](https://github.com/debjitk) for this release
 
 ---
 
@@ -96,7 +121,7 @@ Thanks to [@BabyAzerty](https://github.com/BabyAzerty) for this release
 - Updated SwiftLint to 0.21.0
 
 ### Fixed
-- Now `ConcurrentOperation` is subclassable with `open` instead of `public` Access Control [#2](https://github.com/FabrizioBrancati/Queuer/issue/2)
+- Now `ConcurrentOperation` is subclassable with `open` instead of `public` Access Control [#2](https://github.com/FabrizioBrancati/Queuer/issues/2)
 - Fixed tests that sometimes fails
 
 ---